+++ js-api/src/api/grok_api.g.ts
@@ -321,6 +321,8 @@ export interface IDartApi {
   grok_DockManager_FindNode(m: any, e: any): any;
   grok_DockManager_HandleResize(d: any): any;
   grok_DockManager_OnElementClosed(d: any): any;
+  grok_DockManager_OnPanelShown(d: any): any;
+  grok_DockManager_OnPanelHidden(d: any): any;
   grok_WidgetDescriptor_GetDescriptors(): any;
   grok_WidgetDescriptor_GetByName(name: String): any;
   grok_WidgetDescriptor_Get_Name(d: any): any;
@@ -646,6 +648,8 @@ export interface IDartApi {
   grok_TagElement_Set_Tag(te: any, tag: any): any;
   grok_InputBase_Get_InputType(input: any): any;
   grok_InputBase_Get_DataType(input: any): any;
+  grok_InputBase_Get_ValidationMessages(input: any): any;
+  grok_InputBase_OnValidated(input: any): any;
   grok_InputBase_Get_Property(input: any): any;
   grok_InputBase_Set_Property(input: any, p: any): any;
   grok_InputBase_Get_Root(input: any): any;
@@ -2079,21 +2083,8 @@ export interface IDartApi {
   grok_FuncOptions_Create(): any;
   grok_FuncParamOptions_Create(): any;
 
-  // Generated from ../d4/lib/src/viewer_base/viewer_base.api.g.dart
-  grok_ViewerEvent_Create(): any;
-  grok_ViewerEvent_Get_viewer(x: any): any;
-  grok_ViewerEvent_Set_viewer(x: any, v: any): any;
-  grok_ViewerEvent_Get_type(x: any): any;
-  grok_ViewerEvent_Set_type(x: any, v: String): any;
-  grok_ViewerEvent_Get_eventFlag(x: any): any;
-  grok_ViewerEvent_Set_eventFlag(x: any, v: Bool): any;
-  grok_ViewerEvent_Get_filters(x: any): any;
-  grok_ViewerEvent_Set_filters(x: any, v: any): any;
-  grok_ViewerEvent_Get_row(x: any): any;
-  grok_ViewerEvent_Set_row(x: any, v: Num): any;
-  grok_ViewerEvent_Get_mouseEvent(x: any): any;
-  grok_ViewerEvent_Set_mouseEvent(x: any, v: any): any;
-  grok_ViewerEvent_Get_bitset(x: any): any;
+  // Generated from ../d4/lib/src/common/common.api.g.dart
+  grok_UsageType_Create(): any;
 
   // Generated from ../d4/lib/src/grid/grid.api.g.dart
   grok_GridCellStyle_Create(): any;
@@ -2195,9 +2186,22 @@ export interface IDartApi {
   grok_GridCellStyle_Set_choices(x: any, v: any): any;
   grok_renderMultipleHistograms(g: any, bounds: any, histograms: any, categoryColumn: any, colors: any, tension: Num, normalize: Bool, markerSize: Num, fill: Bool, minBin: Num, maxBin: Num, localMaximum: Bool, highlightedHistogram: Num): any;
 
+  // Generated from ../d4/lib/src/viewer_base/viewer_base.api.g.dart
+  grok_ViewerEvent_Create(): any;
+  grok_ViewerEvent_Get_viewer(x: any): any;
+  grok_ViewerEvent_Set_viewer(x: any, v: any): any;
+  grok_ViewerEvent_Get_type(x: any): any;
+  grok_ViewerEvent_Set_type(x: any, v: String): any;
+  grok_ViewerEvent_Get_eventFlag(x: any): any;
+  grok_ViewerEvent_Set_eventFlag(x: any, v: Bool): any;
+  grok_ViewerEvent_Get_filters(x: any): any;
+  grok_ViewerEvent_Set_filters(x: any, v: any): any;
+  grok_ViewerEvent_Get_row(x: any): any;
+  grok_ViewerEvent_Set_row(x: any, v: Num): any;
+  grok_ViewerEvent_Get_mouseEvent(x: any): any;
+  grok_ViewerEvent_Set_mouseEvent(x: any, v: any): any;
+  grok_ViewerEvent_Get_bitset(x: any): any;
+
   // Generated from ../d4/lib/src/widgets/widgets.api.g.dart
   grok_InputType_Create(): any;
-
-  // Generated from ../d4/lib/src/common/common.api.g.dart
-  grok_UsageType_Create(): any;
 }